+++ tests/drum/run_integration_tests_in_framework_container.sh
@@ -22,7 +22,16 @@ echo "Running pytest:"
 
 pip install pytest pytest-xdist
 
-pytest tests/drum/test_inference_per_framework.py tests/drum/test_fit_per_framework.py --framework-env $1 \
+# install 'pack' package in R env for tests
+if [ "$1" = "r_lang" ]; then
+    Rscript -e "install.packages('pack', Ncpus=4)"
+fi
+
+pytest tests/drum/test_inference_per_framework.py \
+       tests/drum/test_fit_per_framework.py \
+       tests/drum/test_other_cases_per_framework.py \
+       tests/drum/test_unstructured_mode_per_framework.py \
+       --framework-env $1 \
        --junit-xml="$GIT_ROOT/results_integration.xml" \
        -n auto