+++ custom_model_runner/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### [1.9.6] - 2022-07-19
 ##### Changed
 - bumped rpy2>=3.5.2
-- fixed typechecking in R unstructured prediction
+- fixed typechecking in R predictor
 
 #### [1.9.5] - 2022-07-06
 ##### Fixed