+++ custom_model_runner/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+#### [1.9.6] - 2022-07-19
+##### Changed
+- bumped rpy2>=3.5.2
+- fixed typechecking in R predictor
+
 #### [1.9.5] - 2022-07-06
 ##### Fixed
 - Stringify class labels for PMML model